Support keyword-only function parameters
Following an undocumented feature of Skylark-in-Java,
which in turn follows Python3,
a function declaration may now include optional parameters
after the *args parameter:

	def f(a, b, c=1, *args, d=2, **kwargs)

The parameter d is a "keyword-only" parameter as it can
never by assigned from a positional parameter; all positional
arguments surplus to a, b, and c are put in a tuple and
assigned to args.

To declare a non-variadic function with keyword-only arguments,
the *args parameter is replaced by just *:

        def f(a, b, c=1, *, d=2, **kwargs)

The * parameter is not a real parameter; it just serves as a
separator between the parameter that may be specified positionally
and the keyword-only ones.

Fixes #61